requirements.txt
flask==2.0.1
werkzeug==2.0.3
azure-storage-queue==12.1.4
python-dotenv==0.19.0 